Move jabber_app into jabber::app -- making jabber a single library.
