+++ apps/jabber_out.cpp
@@ -4,7 +4,6 @@
  */
 
 #include <jabber.hpp>
-#include <jabber_app.hpp>
 #include <cxxopts.hpp>
 
 #include <iostream>
@@ -13,7 +12,7 @@
 #include <fstream>
 
 using namespace jabber;
-using namespace jabber_app;
+using namespace jabber::app;
 
 int main(int argc, char *argv[])
 {
