+++ apps/jabber_participant.cpp
@@ -9,7 +9,6 @@
  */
 
 #include <jabber.hpp>
-#include <jabber_app.hpp>
 #include <cxxopts.hpp>
 #include <precice/precice.hpp>
 
@@ -28,7 +27,7 @@
 #endif
 
 using namespace jabber;
-using namespace jabber_app;
+using namespace jabber::app;
 
 int main(int argc, char *argv[])
 {
