+++ src/jabber.hpp
@@ -1,0 +1,5 @@
+#include "core/core.hpp"
+
+#ifdef JABBER_WITH_APP
+#include "app/app.hpp"
+#endif // JABBER_WITH_APP
