+++ tests/test_app_config_input.cpp
@@ -7,12 +7,12 @@
 #include <catch2/matchers/catch_matchers_vector.hpp>
 #include <catch2/generators/catch_generators.hpp>
 #include <catch2/generators/catch_generators_range.hpp>
-#include <jabber_app.hpp>
+#include <jabber.hpp>
 
 #include <functional>
 
 using namespace jabber;
-using namespace jabber_app;
+using namespace jabber::app;
 using namespace Catch::Matchers;
 using namespace Catch::Generators;
 namespace jabber_test
