+++ tests/test_compute_3d.cpp
@@ -6,9 +6,6 @@
 #include <catch2/generators/catch_generators_all.hpp>
 
 #include <jabber.hpp>
-#ifdef JABBER_WITH_APP
-#include <jabber_app.hpp>
-#endif // JABBER_WITH_APP
 
 #ifdef JABBER_WITH_OPENMP
 #include <omp.h>
@@ -21,7 +18,7 @@ using namespace jabber;
 using namespace Catch::Matchers;
 using namespace Catch::Generators;
 #ifdef JABBER_WITH_APP
-using namespace jabber_app;
+using namespace jabber::app;
 #endif // JABBER_WITH_APP
 
 namespace jabber_test
